size_t: update ogsf lib
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@52044 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/ogsf.h
@@ -234,7 +234,7 @@ typedef struct
     int data_id;
     int dims[MAXDIMS];
     int ndims;
-    int numbytes;
+    size_t numbytes;
     char *unique_name;
     typbuff databuff;
     IFLAG changed;
